Add `useDrawing` hook for managing the drawing logic Refactor the main `App.tsx` file for a cleaner code structure
